General interface

General interface

Overview and concepts

The general interface supports the following setup for likelihood-based indirect inference, using a structural model $S$ and an auxiliary model $A$, with given data $y$.

For each set of parameters $θ$, a structural model $S$ is used to generate simulated data $x$, ie

\[x_S(θ, ϵ)\]

where $ϵ$ is a set of common random numbers that can be kept constant for various $θ$s. Nevertheless, the above is not necessarily a deterministic relationship, as additional randomness can be used.

An auxiliary model $A$ with parameters $ϕ$ can be estimated using generated data $x$ with maximum likelihood, ie

\[ϕ_A(x) = \arg\max_ϕ p_A(x ∣ ϕ)\]

is the maximum likelihood estimate.

The likelihood of $y$ at parameters $θ$ is obtained as

\[p_A(y ∣ ϕ_A(x_S(θ, ϵ))\]

This is multiplied by a prior in $θ$, specified in logs.

The user should define

types for the structural model and the auxiliary model,

methods for the functions below to dispatch on these types.

component	Julia method

$x_M$	simulate_data

$p_A$	loglikelihood

$ϕ_A$	MLE

draw $ϵ$	common_random

The framework is explained in detail below.

Data

Data can be of any type, since it is generated and used by user-defined functions. Arrays, tuples (optionally named) are recommended for simple models, as there no need to wrap them in a type, as the auxiliary model type is used for dispatch.

More complex data structures may benefit from being wrapped in a struct.

Structural models

A structural model can be used to generate data from parameters $θ$ and random numbers $ϵ$ using simulate_data. The mapping is not necessarily deterministic even given $ϵ$, but the latter can be used for common random numbers for certain models, to make the mapping continuous, or reduce variance in simulations.

When applicable, independent variables (eg covariates) necessary for simulation should be included in the structural model object. It is recommended that a type is defined for each problem, along with methods for simulate_data.

IndirectLikelihood.simulate_data — Function.

simulate_data(rng, structural_model, θ, ϵ)

Simulate data from the structural_model

using random number generator rng,

with parameters θ,

common random numbers ϵ.

Usage

The user should define a method for this function for each structural_model type with the signature

simulate_data(rng::AbstractRNG, structural_model, θ, ϵ)

This should return simulated data in the format that can be used by MLE and loglikelihood.

For infeasible/meaningless parameters, nothing should be returned.

source

simulate_data(rng, structural_model, θ)

Simulate data, generating ϵ using rng.

See common_random.

Usage

For interactive/exploratory use. Models should define methods for simulate_data(rng::AbstractRNG, structural_model, θ, ϵ).

source

simulate_data(structural_model, θ)

Simulate data, generating ϵ with the default random number generator.

See common_random.

Usage

For interactive/exploratory use. Models should define methods for simulate_data(rng::AbstractRNG, structural_model, θ, ϵ).

source

Common random numbers

Common random numbers are a set of random numbers that can be re-used for for simulation with different parameter values.

common_random should yield a random value for the random variables (usually an Array or a collection of similar structures).

IndirectLikelihood.common_random — Function.

common_random(rng, structural_model)

Return common random numbers that can be reused by simulate_data with different parameters.

When the model structure does not allow common random numbers, return nothing.

The first argument is the random number generator.

Usage

The user should define a method for this function for each structural_model type.

See also common_random! for further optimizations.

source

For error structures which can be overwritten in place, the user can define common_random! as an optimization.

IndirectLikelihood.common_random! — Function.

common_random!(rng, structural_model, ϵ)

Update the common random numbers for the model. The semantics is as follows:

it may, but does not need to, change the contents of its second argument,

the new common random numbers should be returned regardless.

Two common usage patterns are

having a mutable ϵ, updating that in place, returning ϵ,

generating new ϵ, returning that.

Note

The default method falls back to common_random, reallocating with each call. A method for this function should be defined only when allocations can be optimized.

source

common_random!(rng, problem)

Return a new problem with updated common random numbers.

source

Likelihood

These methods should be defined for stuctural model types, and accept data from simulate_data.

IndirectLikelihood.MLE — Function.

MLE(auxiliary_model, data)

Maximum likelihood estimate of the parameters for data in model.

When ϕ == MLE(auxiliary_model, data), ϕ should maximize

ϕ -> loglikelihood(auxiliary_model, data, ϕ)

See loglikelihood.

Methods should be defined by the user for each auxiliary_model type. A fallback method is defined for nothing as data (infeasible parameters). See also simulate_data.

source

IndirectLikelihood.loglikelihood — Function.

loglikelihood(auxiliary_model, data, ϕ)

Log likelihood of data under auxiliary_model with parameters $ϕ$. See MLE.

source

Problem framework

IndirectLikelihood.IndirectLikelihoodProblem — Type.

IndirectLikelihoodProblem(data, log_prior, structural_model, auxiliary_model, ϵ)

A simple wrapper for an indirect likelihood problem.

The user should implement simulate_data, MLE, loglikelihood, and common_random.

source

IndirectLikelihood.simulate_problem — Function.

simulate_problem([rng], log_prior, structural_model, auxiliary_model, θ)

Initialize an IndirectLikelihoodProblem with simulated data, using parameters θ.

Useful for debugging and exploration of identification with simulated data.

source

Utilities

IndirectLikelihood.local_jacobian — Function.

local_jacobian(problem, θ₀, ω_to_θ; vecϕ)

Calculate the local Jacobian of the estimated auxiliary parameters $ϕ$ at the structural parameters $θ=θ₀$.

ω_to_θ maps a vector of reals ω to the parameters θ in the format acceptable to structural_model. It should support ContinuousTransformations.transform and ContinuousTransformations.inverse. See, for example, ContinuousTransformations.TransformationTuple.

vecϕ is a function that is used to flatten the auxiliary parameters to a vector. Defaults to vec_parameters.

source

IndirectLikelihood.vec_parameters — Function.

Return the values of the argument as a vector, potentially (but not necessarily) restricting to those elements that uniquely determine the argument.

For example, a symmetric matrix would be determined by the diagonal and either half.

source

IndirectLikelihood.indirect_loglikelihood — Function.

indirect_loglikelihood(auxiliary_model, y, x)

estimate auxiliary_model from summary statistics x using maximum likelihood,

return the likelihood of summary statistics y under the estimated parameters.

Useful for pseudo-likelihood indirect inference, where y would be the observed and x the simulated data.

source